#24 some tokenizer related refactoring
